requirements.txt
# UI自动化核心
selenium>=4.18.1
webdriver-manager>=4.0.1  # 自动管理浏览器驱动（无需手动下载）
# 测试框架
pytest>=7.4.3
pytest-rerunfailures>=12.0  # 用例失败重跑
pytest-xdist>=3.5.0         # 多线程执行
# 报告
pytest-html>=3.2.0
allure-pytest>=2.13.2       # Allure报告（企业级首选）
# 配置与数据
PyYAML>=6.0.1               # YAML配置读取（比INI/CSV灵活）
openpyxl>=3.1.2             # Excel数据读取
jsonpath-ng>=1.6.0          # JSON数据解析
# 安全与工具
cryptography>=42.0.0        # 敏感配置加密/解密
loguru>=0.7.2               # 增强日志（替代原生logging）
pillow>=10.2.0              # 截图处理
colorlog>=6.8.2             # 彩色日志
python-dotenv>=1.0.1        # 环境变量读取
# 邮件推送
yagmail>=0.15.271           # 简单易用的邮件库
# 跨平台兼容
pathlib>=1.0.1              # 路径处理